Reorganize huge data chunk into a separate file
Google scripts allows for separate files that all load automatically,
so we can utilize this to clean up the code a bit. This commit moves
the huge images object block to a separate 'data' file and then
loads it in the main 'code' through the variable.

This requires us to open a new script file in Google Script editor
('File' -> 'New' -> 'Script file') and name it 'data', then paste
the code from data.js into it.

This separation of files can also help us make the code a bit more
sustaining for the future, in case this experiment continues, without
having to mess with a humongous block of text to fix something small
or add more images, etc.